Add validation and patientHistory resolvers

+++ src/main/kotlin/com/mpralat/patienthistory/graphql/DoctorMutationResolver.kt
@@ -5,6 +5,7 @@ import com.mpralat.patienthistory.domain.IdProvider
 import com.mpralat.patienthistory.domain.doctor.Doctor
 import com.mpralat.patienthistory.domain.doctor.Specialty
 import com.mpralat.patienthistory.domain.doctor.repository.DoctorRepository
+import com.mpralat.patienthistory.graphql.exception.CustomGraphQLException
 import org.springframework.stereotype.Component
 
 @Component
@@ -24,11 +25,13 @@ class DoctorMutationResolver(
     }
 
     fun deleteDoctor(id: String): Boolean {
+        if (!doctorRepository.existsById(id)) throw CustomGraphQLException("No doctor with id: $id")
         doctorRepository.deleteById(id)
         return true
     }
 
     fun updateDoctor(id: String, firstName: String?, lastName: String?): Doctor? {
+        if (!doctorRepository.existsById(id)) throw CustomGraphQLException("No doctor with id: $id")
         return doctorRepository.updateDoctor(
                 id,
                 firstName,
@@ -37,6 +40,7 @@ class DoctorMutationResolver(
     }
 
     fun addSpecialty(id: String, newSpecialties: List<Specialty>): Doctor? {
+        if (!doctorRepository.existsById(id)) throw CustomGraphQLException("No doctor with id: $id")
         return doctorRepository.addSpecialties(
                 id = id,
                 newSpecialties = newSpecialties
